+++ app/controllers/images_controller.rb
@@ -126,9 +126,10 @@ class ImagesController < ApplicationController
 
   # GET /images/download
   def download
-    send_data(Export::Download.generate_csv(Image.where(project_id: sessions_current_project_id)),
-              type: 'text',
-              filename: "images_#{DateTime.now}.csv")
+    send_data(
+      Export::Download.generate_csv(Image.where(project_id: sessions_current_project_id)),
+      type: 'text',
+      filename: "images_#{DateTime.now}.csv")
   end
 
   # GET /images/:id/extract/:x/:y/:height/:width