+++ app/helpers/observation_matrices/export/nexml_helper.rb
@@ -21,16 +21,10 @@ module ObservationMatrices::Export::NexmlHelper
                 if cs.depictions.load.any?
                   xml.state(id: "cs#{cs.id}", label: cs.target_name(:key, nil), symbol: "#{i}") do
                     cs.depictions.each do |d|
-                      if d.image.image_file_content_type == 'image/tiff'
-                        href = d.image.image_file.url(:medium)
-                      else
-                        href = d.image.image_file.url(:original)
-                      end
-
                       xml.meta(
                         'xsi:type' => 'ResourceMeta',
                         'rel' => 'foaf:depiction',
-                        'href' => short_url(href), # root_url + d.image.image_file.url(:original)[1..-1]
+                        'href' => image_short_url(d.image), # see app/helpers/images_helper.rb
                         'label' => c.target_name(:description, nil) + ' ' + cs.target_name(:description, nil)
                       )
                     end
@@ -88,8 +82,8 @@ module ObservationMatrices::Export::NexmlHelper
         end # end format
 
         include_continuous_matrix(opt.merge(descriptors: d)) if opt[:include_matrix]
-    end # end multistate characters
-    opt[:target]
+      end # end multistate characters
+      opt[:target]
   end
 
   def include_multistate_matrix(options = {})
@@ -100,17 +94,21 @@ module ObservationMatrices::Export::NexmlHelper
 
     # the matrix
     cells = m.observations_in_grid({})[:grid]
+
     p = m.observation_matrix_columns.order('observation_matrix_columns.position').map(&:descriptor_id)
-    q = m.observation_matrix_rows.order('observation_matrix_rows.position').collect{|i| "#{i.otu_id}|#{i.collection_object_id}" }
+
+    q = m.observation_matrix_rows.order('observation_matrix_rows.position').collect{|i| i.observation_index } # could pluck this string concat from the db
+
     xml.matrix do
       m.observation_matrix_rows.each do |r|
-        xml.row(id: "multistate_row#{r.id}", otu: "row_#{r.id}") do |row| # use row_id to uniquel identify the row ##  Otu#id to uniquely id the row
+        xml.row(id: "multistate_row#{r.id}", otu: "row_#{r.id}") do |row| # use row_id to uniquely identify the row, Otu#id to uniquely id the row Otu
 
           # cell representation
           opt[:descriptors].each do |d|
 
             x = p.index(d.id) # opt[:descriptors].index(d)  #   .index(d)
-            y = q.index("#{r.otu_id}|#{r.collection_object_id}")
+            y = q.index( r.observation_index)
+
             observations = cells[ x ][ y ]
 
             case observations.size
@@ -151,7 +149,7 @@ module ObservationMatrices::Export::NexmlHelper
     # the matrix
     cells = m.observations_in_grid({})[:grid]
 
-    z = m.observation_matrix_rows.map.collect{|i| "#{i.otu_id}|#{i.collection_object_id}"}
+    z = m.observation_matrix_rows.collect{|i| i.observation_index }
 
     xml.matrix do |mx|
       m.observation_matrix_rows.each do |o|
@@ -161,7 +159,7 @@ module ObservationMatrices::Export::NexmlHelper
           opt[:descriptors].each do |c|
 
             x = m.descriptors.index(c)
-            y = z.index("#{o.otu_id}|#{o.collection_object_id}")
+            y = z.index( o.observation_index )
 
             observations = cells[ x ][ y ]
             if observations.size > 0  && !observations.first.continuous_value.nil?
@@ -173,8 +171,6 @@ module ObservationMatrices::Export::NexmlHelper
     end # end matrix
     return opt[:target]
   end
-
-
 
   def nexml_otus(options = {})
     opt = {target: ''}.merge!(options)
@@ -191,12 +187,14 @@ module ObservationMatrices::Export::NexmlHelper
           about: "#row_#{r.id}", # technically only need this for proper RDFa extraction  !!! Might need this to be different, is it about row, or row object!
           label: observation_matrix_row_label_nexml(r)
         ) do
-          include_collection_objects(opt.merge(otu: r.row_object)) if opt[:include_collection_objects]
+          # TODO: should otu: be more generic
+          include_collection_objects(opt.merge(otu: r.observation_object)) if opt[:include_collection_objects]
+
           # This is experimental only. Issues:
           # * It draws images and data from all matrices, # not just this one
           # * Depictions are on Observation::Media, not OTU, i.e. we could be more granular throughout
           # * Citations are on Image, not Depiction
-          im =  ImageMatrix.new(project_id: r.project_id, otu_filter: r.otu_id.to_s)
+          im = Tools::ImageMatrix.new(project_id: r.project_id, otu_filter: r.otu_id.to_s)
           descriptors = im.list_of_descriptors.values
           if !im.blank? && !im.depiction_matrix.empty?
             object = im.depiction_matrix.first
@@ -214,27 +212,20 @@ module ObservationMatrices::Export::NexmlHelper
                 lbl.push(attribution_nexml_label(img_attr)) unless img_attr.nil?
                 lbl = lbl.compact.join('; ')
 
-                if im.image_hash[depiction[:image_id]][:image_file_content_type] == 'image/tiff'
-                  href = short_url(im.image_hash[depiction[:image_id]][:medium_url])
-                else
-                  href = short_url(im.image_hash[depiction[:image_id]][:original_url])
-                end
-
                 xml.meta(
                   'xsi:type' => 'ResourceMeta',
                   'rel' => 'foaf:depiction',
                   'about' => "row_#{r.id}",
-                  'href' => href,
+                  'href' => image_short_url( im.image_hash[depiction[:image_id]][:original_url] ),
+                  'label' => lbl
                   #'object' => object[1][:object].otu_name,
                   #'label' => descriptors[index][:name], ###
-                  'label' => lbl
                   #'caption' => depiction[:caption],
                   #'citation' => im.image_hash[depiction[:image_id]][:citations].collect{|i| i[:cached]}.join('')
                 )
               end
             end
           end
-
 
 =begin
           Observation::Media.where(otu_id: r.otu_id).each do |o|
@@ -255,8 +246,6 @@ module ObservationMatrices::Export::NexmlHelper
             end
           end
 =end
-
-
         end
       end
     end
@@ -283,40 +272,40 @@ module ObservationMatrices::Export::NexmlHelper
     return opt[:target]
   end
 
-   #
-   # NOT USED - current using inline approach to handle depictions
+  #
+  # NOT USED - current using inline approach to handle depictions
   def nexml_otu_depictions(options = {})
-     opt = {target:  '', descriptors: []}.merge!(options)
-     xml = Builder::XmlMarkup.new(target: opt[:target])
-     m = opt[:observation_matrix]
-     otus = m.otus
-     im =  ImageMatrix.new(
-       project_id: m.project_id,
-       otu_filter: m.otus.pluck(:id).join('|'))
-     descriptors = im.list_of_descriptors.values
-     row_hash = m.observation_matrix_rows.map{|i| [i.otu_id, i.id]}.to_h
+    opt = {target:  '', descriptors: []}.merge!(options)
+    xml = Builder::XmlMarkup.new(target: opt[:target])
+    m = opt[:observation_matrix]
+    otus = m.otus
+    im = Tools::ImageMatrix.new(
+      project_id: m.project_id,
+      otu_filter: m.otus.pluck(:id).join('|'))
+    descriptors = im.list_of_descriptors.values
+    row_hash = m.observation_matrix_rows.map{|i| [i.otu_id, i.id]}.to_h
 
-     xml.otu_depictions do |d|
-       im.depiction_matrix.each do |object|
-         object[1][:depictions].each_with_index do |depictions, index|
-           depictions.each do |depiction|
+    xml.otu_depictions do |d|
+      im.depiction_matrix.each do |object|
+        object[1][:depictions].each_with_index do |depictions, index|
+          depictions.each do |depiction|
 
-             xml.meta(
-               'xsi:type' => 'ResourceMeta',
-               'rel' => 'foaf:depiction',
-               'about' => "row_#{row_hash[object[1][:otu_id].to_i].to_s}",
-               'href' => root_url + im.image_hash[depiction[:image_id]][:original_url],
-               'object' => object[1][:object].otu_name,
-               'description' => descriptors[index][:caption],
-               'label' => depiction[:figure_label],
-               'citation' => depiction[:source_cached]
-             )
-           end
-         end
-       end
-     end
-     return opt[:target]
-   end
+            xml.meta(
+              'xsi:type' => 'ResourceMeta',
+              'rel' => 'foaf:depiction',
+              'about' => "row_#{row_hash[object[1][:otu_id].to_i].to_s}",
+              'href' => root_url + im.image_hash[depiction[:image_id]][:original_url],
+              'object' => object[1][:object].otu_name,
+              'description' => descriptors[index][:caption],
+              'label' => depiction[:figure_label],
+              'citation' => depiction[:source_cached]
+            )
+          end
+        end
+      end
+    end
+    return opt[:target]
+  end
 
 
   # TODO: if semantics change we can add them as block later.