+++ app/helpers/observation_matrices_helper.rb
@@ -41,14 +41,14 @@ module ObservationMatricesHelper
 
   # @return [String]
   #   the fully formatted cell, handles polymorphisms
-  #   show states in tnt or nexus format for a 'cell' (e.g. [ab]) 
+  #   show states in tnt or nexus format for a 'cell' (e.g. [ab])
   #   Mx.print_codings in mx
-  def observations_cell_label(observations_hash, descriptor, otu_co, style = :tnt)
-    case observations_hash[descriptor.id][otu_co].size
+  def observations_cell_label(observations_hash, descriptor, hash_index, style = :tnt)
+    case observations_hash[descriptor.id][hash_index].size
     when 0
       "?"
     when 1
-      o = observations_hash[descriptor.id][otu_co][0]
+      o = observations_hash[descriptor.id][hash_index][0]
       s = observation_export_value(o)
 
       if s.length > 1 && style == :nexus && o.type == 'Observation::Qualitative'
@@ -57,12 +57,15 @@ module ObservationMatricesHelper
         s
       end
     else
-      str = observations_hash[descriptor.id][otu_co].collect{|o| observation_export_value(o) }.sort.join("")
+      str = observations_hash[descriptor.id][hash_index].collect{|o| observation_export_value(o) }.sort
+
       case style
+      when :csv
+        str.join('|')
       when :nexus
-        "{#{str}}"
+        "{#{str.join("")}}"
       else
-        "[#{str}]"
+        "[#{str.join("")}]"
       end
     end
   end
@@ -73,7 +76,7 @@ module ObservationMatricesHelper
     case observation.type
     when 'Observation::Qualitative'
       observation.character_state.label
-    when 'Observation::PresenceAbsence' 
+    when 'Observation::PresenceAbsence'
       case observation.presence
       when true
         '1'
@@ -87,8 +90,22 @@ module ObservationMatricesHelper
     when 'Observation::Continuous'
       observation.converted_value.to_s
     else
-      '-' # ? not sure 
+      '-' # ? not sure
     end
+  end
+
+  def descriptor_list(observation_matrix)
+    rows = []
+    observation_matrix.descriptors.order(:position).each do |d|
+      l = d.name + ': '
+      if d.qualitative?
+        l << d.character_states.order(:position).collect{|cs| "#{cs.label}: " + cs.name }.join('; ')
+      elsif d.presence_absence?
+        l << '0: absent; 1: present'
+      end
+      rows.push l
+    end
+    rows.join("\n")
   end
 
 end