+++ app/helpers/observation_matrix_rows_helper.rb
@@ -3,7 +3,7 @@ module ObservationMatrixRowsHelper
   # Display in app
   def observation_matrix_row_tag(observation_matrix_row)
     return nil if observation_matrix_row.nil?
-    object_tag(observation_matrix_row.row_object)
+    object_tag(observation_matrix_row.observation_object)
   end
 
   def observation_matrix_row_link(observation_matrix_row)
@@ -17,7 +17,7 @@ module ObservationMatrixRowsHelper
   # TODO: remove for pattern below/in helpers README
   def observation_matrix_row_label(observation_matrix_row)
     return observation_matrix_row.name unless observation_matrix_row.name.blank?
-    o = observation_matrix_row.row_object
+    o = observation_matrix_row.observation_object
     s = label_for(o) 
     s.gsub!(/[^\w]/, '_')
     s[0..11] + "_#{o.id}"
@@ -27,14 +27,14 @@ module ObservationMatrixRowsHelper
 
   def observation_matrix_row_label_nexus(observation_matrix_row)
     return observation_matrix_row.name unless observation_matrix_row.name.blank?
-    o = observation_matrix_row.row_object
+    o = observation_matrix_row.observation_object
     s = label_for(o)
     s.gsub!(/[^\w]/, '_')
   end
 
   def observation_matrix_row_label_nexml(observation_matrix_row)
     return observation_matrix_row.name unless observation_matrix_row.name.blank?
-    o = observation_matrix_row.row_object
+    o = observation_matrix_row.observation_object
     s = label_for(o)
   end