+++ app/javascript/vue/components/radials/object/helpers/makeEmptyCitation.js
@@ -1,0 +1,6 @@
+export default () => ({
+  id: undefined,
+  source_id: undefined,
+  pages: undefined,
+  is_original: undefined
+})