+++ app/javascript/vue/tasks/observation_matrices/matrix_row_coder/App.vue
@@ -26,6 +26,8 @@ import NavigationMatrix from './MatrixRowCoder/Navigation/NavigationMatrix.vue'
 import { ObservationMatrix } from 'routes/endpoints'
 
 export default {
+  name: 'MatrixRowCoderApp',
+
   components: {
     MatrixRowCoder,
     NavigationMatrix,