+++ app/javascript/vue/tasks/observation_matrices/matrix_row_coder/store/mutations/setObservation.js
@@ -9,8 +9,34 @@ export default function (state, observation) {
 
   if (observation.type === ObservationTypes.Qualitative && !observation.characterStateId) { throw `Qualitative Observations must have a character state id!` }
 
+  const observations = state.observations.filter(o => o.descriptorId === observation.descriptorId)
   let existingObservation
-  if (observation.type !== ObservationTypes.Qualitative) { existingObservation = state.observations.find(o => o.descriptorId === observation.descriptorId) } else { existingObservation = state.observations.find(o => o.descriptorId === observation.descriptorId && o.characterStateId === observation.characterStateId) }
 
-  if (existingObservation) { Object.assign(existingObservation, observation) } else { state.observations.push(observation) }
-};
+  if (observation.type === ObservationTypes.FreeText) {
+    existingObservation = observations.find(o => o.descriptorId === observation.descriptorId)
+  } else if (observation.type === ObservationTypes.Qualitative) {
+    existingObservation = observations.find(o => o.characterStateId === observation.characterStateId)
+  } else {
+    existingObservation = getObservation(observation)
+  }
+
+  if (existingObservation) {
+    Object.assign(existingObservation, observation)
+  } else {
+    state.observations.push(observation)
+  }
+
+  function getObservation (observation) {
+    if (observations.length) {
+      const o = observations.find(o => o.id === observation.id)
+
+      if (o) { return o }
+
+      if (observations.length === 1) {
+        return !observations[0].id && observations[0]
+      }
+    }
+
+    return false
+  }
+}