+++ app/javascript/vue/tasks/observation_matrices/new/store/actions/loadRowItems.js
@@ -5,7 +5,7 @@ import getPagination from 'helpers/getPagination.js'
 export default function ({ commit, state }, params = {}) {
   return new Promise((resolve, reject) => {
     state.settings.loadingRows = true
-    return GetMatrixObservationRows(state.matrix.id, Object.assign({}, params, state.configParams)).then(response => {
+    return GetMatrixObservationRows(state.matrix.id, Object.assign({}, params, state.configParams, { extend: ['observation_object'] })).then(response => {
       commit(MutationNames.SetMatrixRows, response.body)
       commit(MutationNames.SetRowFixedPagination, getPagination(response))
       state.settings.loadingRows = false