+++ app/javascript/vue/tasks/uniquify/people/components/tableRoles.vue
@@ -5,29 +5,33 @@
       <tbody>
         <tr
           v-for="(item, index) in list"
+          :key="item.id"
           class="contextMenuCells"
-          :class="{ even: (index % 2 == 0) }">
-          <td class="column-property">{{ item.role_object_type }}</td>
-          <td class="column-person">{{ item.role_object_tag }}</td>
+          :class="{ even: (index % 2 === 0) }"
+        >
+          <td class="column-property">
+            {{ item.role_object_type }}
+          </td>
+          <td class="column-person">
+            {{ item.object_tag }}
+          </td>
         </tr>
       </tbody>
     </table>
   </div>
 </template>
 
-<script>
-export default {
-  props: {
-    list: {
-      type: Array,
-      default: () => { return [] }
-    },
-    title: {
-      type: String,
-      required: true
-    }
+<script setup>
+const props = defineProps({
+  list: {
+    type: Array,
+    default: () => []
+  },
+  title: {
+    type: String,
+    required: true
   }
-}
+})
 </script>
 
 <style scoped>